+++ core/risk/engine.go
@@ -71,21 +71,23 @@ type marginChange struct {
 // Engine is the risk engine.
 type Engine struct {
 	Config
-	marginCalculator       *types.MarginCalculator
-	scalingFactorsUint     *scalingFactorsUint
-	log                    *logging.Logger
-	cfgMu                  sync.Mutex
-	model                  Model
-	factors                *types.RiskFactor
-	waiting                bool
-	ob                     Orderbook
-	as                     AuctionState
-	timeSvc                TimeService
-	broker                 Broker
-	riskFactorsInitialised bool
-	mktID                  string
-	asset                  string
-	positionFactor         num.Decimal
+	marginCalculator        *types.MarginCalculator
+	scalingFactorsUint      *scalingFactorsUint
+	log                     *logging.Logger
+	cfgMu                   sync.Mutex
+	model                   Model
+	factors                 *types.RiskFactor
+	waiting                 bool
+	ob                      Orderbook
+	as                      AuctionState
+	timeSvc                 TimeService
+	broker                  Broker
+	riskFactorsInitialised  bool
+	mktID                   string
+	asset                   string
+	positionFactor          num.Decimal
+	linearSlippageFactor    num.Decimal
+	quadraticSlippageFactor num.Decimal
 }
 
 // NewEngine instantiate a new risk engine.
@@ -103,6 +105,8 @@ func NewEngine(log *logging.Logger,
 	positionFactor num.Decimal,
 	riskFactorsInitialised bool,
 	initialisedRiskFactors *types.RiskFactor, // if restored from snapshot, will be nil otherwise
+	linearSlippageFactor num.Decimal,
+	quadraticSlippageFactor num.Decimal,
 ) *Engine {
 	// setup logger
 	log = log.Named(namedLogger)
@@ -110,21 +114,23 @@ func NewEngine(log *logging.Logger,
 
 	sfUint := scalingFactorsUintFromDecimals(marginCalculator.ScalingFactors)
 	e := &Engine{
-		log:                    log,
-		Config:                 config,
-		marginCalculator:       marginCalculator,
-		model:                  model,
-		waiting:                false,
-		ob:                     ob,
-		as:                     as,
-		timeSvc:                timeSvc,
-		broker:                 broker,
-		mktID:                  mktID,
-		asset:                  asset,
-		scalingFactorsUint:     sfUint,
-		factors:                model.DefaultRiskFactors(),
-		riskFactorsInitialised: riskFactorsInitialised,
-		positionFactor:         positionFactor,
+		log:                     log,
+		Config:                  config,
+		marginCalculator:        marginCalculator,
+		model:                   model,
+		waiting:                 false,
+		ob:                      ob,
+		as:                      as,
+		timeSvc:                 timeSvc,
+		broker:                  broker,
+		mktID:                   mktID,
+		asset:                   asset,
+		scalingFactorsUint:      sfUint,
+		factors:                 model.DefaultRiskFactors(),
+		riskFactorsInitialised:  riskFactorsInitialised,
+		positionFactor:          positionFactor,
+		linearSlippageFactor:    linearSlippageFactor,
+		quadraticSlippageFactor: quadraticSlippageFactor,
 	}
 	stateVarEngine.RegisterStateVariable(asset, mktID, RiskFactorStateVarName, FactorConverter{}, e.startRiskFactorsCalculation, []statevar.EventType{statevar.EventTypeMarketEnactment, statevar.EventTypeMarketUpdated}, e.updateRiskFactor)